+++ dflib/src/main/java/com/nhl/dflib/DataFrameByRowBuilder.java
@@ -40,6 +40,6 @@ public class DataFrameByRowBuilder {
             series[i] = columnBuilders[i].toSeries();
         }
 
-        return new ColumnDataFrame(columnsIndex, series);
+        return new ColumnDataFrame(null, columnsIndex, series);
     }
 }