+++ dflib/src/main/java/com/nhl/dflib/builder/DataFrameFoldByRowBuilder.java
@@ -46,7 +46,7 @@ public class DataFrameFoldByRowBuilder extends BaseDataFrameBuilder {
             series[i] = new ArraySeries(columnarData[i]);
         }
 
-        return new ColumnDataFrame(columnsIndex, series);
+        return new ColumnDataFrame(null, columnsIndex, series);
     }
 
     public <T> DataFrame ofIterable(Iterable<T> iterable) {
@@ -87,7 +87,7 @@ public class DataFrameFoldByRowBuilder extends BaseDataFrameBuilder {
             series[i] = new DoubleArraySeries(columnarData[i]);
         }
 
-        return new ColumnDataFrame(columnsIndex, series);
+        return new ColumnDataFrame(null, columnsIndex, series);
     }
 
     public DataFrame ofStream(DoubleStream stream) {
@@ -127,7 +127,7 @@ public class DataFrameFoldByRowBuilder extends BaseDataFrameBuilder {
             series[i] = new IntArraySeries(columnarData[i]);
         }
 
-        return new ColumnDataFrame(columnsIndex, series);
+        return new ColumnDataFrame(null, columnsIndex, series);
     }
 
     public DataFrame ofStream(IntStream stream) {
@@ -167,7 +167,7 @@ public class DataFrameFoldByRowBuilder extends BaseDataFrameBuilder {
             series[i] = new LongArraySeries(columnarData[i]);
         }
 
-        return new ColumnDataFrame(columnsIndex, series);
+        return new ColumnDataFrame(null, columnsIndex, series);
     }
 
     public DataFrame ofStream(LongStream stream) {