+++ dflib/src/test/java/com/nhl/dflib/DataFrame_HashJoinsTest.java
@@ -4,54 +4,10 @@ import com.nhl.dflib.join.JoinIndicator;
 import com.nhl.dflib.unit.DataFrameAsserts;
 import org.junit.jupiter.api.Test;
 
-import java.util.Objects;
-
-public class DataFrame_JoinsTest {
+public class DataFrame_HashJoinsTest {
 
     @Test
-    public void nestedLoop_Inner() {
-
-        DataFrame df1 = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y");
-
-        DataFrame df2 = DataFrame.foldByRow("c", "d").of(
-                2, "a",
-                2, "b",
-                3, "c");
-
-        DataFrame df = df1.innerJoin()
-                .predicatedBy((lr, rr) -> Objects.equals(lr.get(0), rr.get(0)))
-                .with(df2);
-
-        new DataFrameAsserts(df, "a", "b", "c", "d")
-                .expectHeight(2)
-                .expectRow(0, 2, "y", 2, "a")
-                .expectRow(1, 2, "y", 2, "b");
-    }
-
-    @Test
-    public void nestedLoop_Inner_NoMatches() {
-
-        DataFrame df1 = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y");
-
-        DataFrame df2 = DataFrame.foldByRow("c", "d").of(
-                2, "a",
-                2, "b",
-                3, "c");
-
-        DataFrame df = df1.innerJoin()
-                .predicatedBy((lr, rr) -> false)
-                .with(df2);
-
-        new DataFrameAsserts(df, "a", "b", "c", "d")
-                .expectHeight(0);
-    }
-
-    @Test
-    public void nestedLoop_Inner_IndexOverlap() {
+    public void asRight() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -63,115 +19,61 @@ public class DataFrame_JoinsTest {
                 3, "c");
 
         DataFrame df = df1.innerJoin()
-                .predicatedBy((lr, rr) -> Objects.equals(lr.get(0), rr.get(0)))
-                .with(df2);
+                .on(0)
+                .with(df2.as("df2"));
 
-        new DataFrameAsserts(df, "a", "b", "a_", "b_")
+        new DataFrameAsserts(df, "a", "b", "df2.a", "df2.b")
                 .expectHeight(2)
                 .expectRow(0, 2, "y", 2, "a")
                 .expectRow(1, 2, "y", 2, "b");
     }
 
     @Test
-    public void nestedLoop_Left() {
+    public void asLeft() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
                 2, "y");
 
-        DataFrame df2 = DataFrame.foldByRow("c", "d").of(
+        DataFrame df2 = DataFrame.foldByRow("a", "b").of(
                 2, "a",
                 2, "b",
                 3, "c");
 
-        DataFrame df = df1
-                .leftJoin()
-                .predicatedBy((lr, rr) -> Objects.equals(lr.get(0), rr.get(0)))
+        DataFrame df = df1.as("df1").innerJoin()
+                .on(0)
                 .with(df2);
 
-        new DataFrameAsserts(df, "a", "b", "c", "d")
-                .expectHeight(3)
-                .expectRow(0, 1, "x", null, null)
-                .expectRow(1, 2, "y", 2, "a")
-                .expectRow(2, 2, "y", 2, "b");
+        new DataFrameAsserts(df, "df1.a", "df1.b", "a", "b")
+                .expectHeight(2)
+                .expectRow(0, 2, "y", 2, "a")
+                .expectRow(1, 2, "y", 2, "b");
     }
 
     @Test
-    public void nestedLoop_Right() {
+    public void asBoth() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
                 2, "y");
 
-        DataFrame df2 = DataFrame.foldByRow("c", "d").of(
+        DataFrame df2 = DataFrame.foldByRow("a", "b").of(
                 2, "a",
                 2, "b",
                 3, "c");
 
-        DataFrame df = df2
-                .rightJoin()
-                .predicatedBy((lr, rr) -> Objects.equals(lr.get(0), rr.get(0)))
-                .with(df1);
+        DataFrame df = df1.as("df1").innerJoin()
+                .on(0)
+                .with(df2.as("df2"));
 
-        new DataFrameAsserts(df, "c", "d", "a", "b")
-                .expectHeight(3)
-                .expectRow(0, null, null, 1, "x")
-                .expectRow(1, 2, "a", 2, "y")
-                .expectRow(2, 2, "b", 2, "y");
+        new DataFrameAsserts(df, "df1.a", "df1.b", "df2.a", "df2.b")
+                .expectHeight(2)
+                .expectRow(0, 2, "y", 2, "a")
+                .expectRow(1, 2, "y", 2, "b");
     }
 
     @Test
-    public void nestedLoop_Full() {
-
-        DataFrame df1 = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y");
-
-        DataFrame df2 = DataFrame.foldByRow("c", "d").of(
-                2, "a",
-                2, "b",
-                3, "c");
-
-        DataFrame df = df1
-                .fullJoin()
-                .predicatedBy((lr, rr) -> Objects.equals(lr.get(0), rr.get(0)))
-                .with(df2);
-
-        new DataFrameAsserts(df, "a", "b", "c", "d")
-                .expectHeight(4)
-                .expectRow(0, 1, "x", null, null)
-                .expectRow(1, 2, "y", 2, "a")
-                .expectRow(2, 2, "y", 2, "b")
-                .expectRow(3, null, null, 3, "c");
-    }
-
-    @Test
-    public void nestedLoop_Indicator() {
-
-        DataFrame df1 = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y");
-
-        DataFrame df2 = DataFrame.foldByRow("c", "d").of(
-                2, "a",
-                2, "b",
-                3, "c");
-
-        DataFrame df = df1.fullJoin()
-                .predicatedBy((lr, rr) -> Objects.equals(lr.get(0), rr.get(0)))
-                .indicatorColumn("ind")
-                .with(df2);
-
-        new DataFrameAsserts(df, "a", "b", "c", "d", "ind")
-                .expectHeight(4)
-                .expectRow(0, 1, "x", null, null, JoinIndicator.left_only)
-                .expectRow(1, 2, "y", 2, "a", JoinIndicator.both)
-                .expectRow(2, 2, "y", 2, "b", JoinIndicator.both)
-                .expectRow(3, null, null, 3, "c", JoinIndicator.right_only);
-    }
-
-    @Test
-    public void hash_Inner() {
+    public void inner() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -196,7 +98,7 @@ public class DataFrame_JoinsTest {
     }
 
     @Test
-    public void hash_Full_IntColumn() {
+    public void full_IntColumn() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -222,7 +124,7 @@ public class DataFrame_JoinsTest {
 
 
     @Test
-    public void hash_Inner_Indexed_HashOverlap() {
+    public void inner_Indexed_HashOverlap() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -244,7 +146,7 @@ public class DataFrame_JoinsTest {
     }
 
     @Test
-    public void hash_Left() {
+    public void left() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -267,7 +169,7 @@ public class DataFrame_JoinsTest {
     }
 
     @Test
-    public void hash_Right_ByPos() {
+    public void right_ByPos() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -290,7 +192,7 @@ public class DataFrame_JoinsTest {
     }
 
     @Test
-    public void hash_Right_ByName() {
+    public void right_ByName() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -313,7 +215,7 @@ public class DataFrame_JoinsTest {
     }
 
     @Test
-    public void hash_Right_ByMatchingName() {
+    public void right_ByMatchingName() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -336,7 +238,7 @@ public class DataFrame_JoinsTest {
     }
 
     @Test
-    public void hash_Full() {
+    public void full() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -360,7 +262,7 @@ public class DataFrame_JoinsTest {
     }
 
     @Test
-    public void hash_MultiColumnHash() {
+    public void multiColumnHash() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -383,7 +285,7 @@ public class DataFrame_JoinsTest {
     }
 
     @Test
-    public void hash_Indicator() {
+    public void indicator() {
 
         DataFrame df1 = DataFrame.foldByRow("a", "b").of(
                 1, "x",